Day 2

+++ aoc_2023/src/day_1/part_1.rs
@@ -1,9 +1,9 @@
-fn process_line(line: &str) -> usize {
-  let numbers: Vec<usize> = line
+fn process_line(line: &str) -> u32 {
+  let numbers: Vec<u32> = line
     .chars()
     .into_iter()
     .fold(Vec::new(), |mut acc, char| {
-      match char.to_string().parse::<usize>() {
+      match char.to_string().parse::<u32>() {
         Ok(char_string_as_uint) => {
           acc.push(char_string_as_uint);
           acc
@@ -27,7 +27,7 @@ fn process_line(line: &str) -> usize {
     let result_as_string = format!("{}{}", first, last);
 
     // I can probably assume unwrapping is okay here
-    result_as_string.parse::<usize>().unwrap()
+    result_as_string.parse::<u32>().unwrap()
   } else {
     0
   }
@@ -36,7 +36,7 @@ fn process_line(line: &str) -> usize {
 // part 1 too low: 36819
 // part 1 too low: 37078
 // part 1 correct: 55123
-pub fn run(contents: &str) -> usize {
+pub fn run(contents: &str) -> u32 {
   contents
     .lines()
     .into_iter()
